app/requirements.txt
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mysql-connector-python==9.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.2
plotly==5.24.1
prompt-toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
pyparsing==3.1.2
pytelegrambotapi==4.22.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
robust==0.2.7
rpds-py==0.20.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.1
telebot==0.0.5
tenacity==9.0.0
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
zipp==3.19.2
